requirements.txt
fastapi
uvicorn[standard]
spotdl
yt-dlp
azure-storage-blob
openai
youtube-transcript-api
